Implement Adding function
